Improve naming and split tests
